+++ src/obfuscationMethods/utils.ts
@@ -1,0 +1,17 @@
+export const shuffle = (array: any[]) => {
+  let currentIndex = array.length;
+  let randomIndex = 0;
+
+  // While there remain elements to shuffle
+  while (currentIndex > 0) {
+    randomIndex = Math.floor(Math.random() * currentIndex);
+    currentIndex--;
+
+    [array[currentIndex], array[randomIndex]] = [
+      array[randomIndex],
+      array[currentIndex],
+    ];
+  }
+
+  return array;
+}
